shell: moved `dl` to it's own file and added support for directories
Both `archive/tar` and `archive/zip` have issues storing subdirectories.
Weirdly this only happens if not using compression, otherwise it works fine.

I'm not sure if this is user error or an issue in the stdlib, but as a
work-around I've elected to always compress when downloading directories.
